Reduced code duplicacy in testing suite
